Import C++ modernization changes from Gerhard Raven.
There are compilation problems, specifically private copy constructors in
EvtItgPtrFunction and related classes, give errors when auto and make_unique
are used in EvtBtoXsgammaKagan, for example. Similar issues occur for
EvtRareLbToLll, EvtVubBLNP and EvtVubNLO.

Not added to this repository yet is the LHCb-only EvtBsMuMuKK model. Need to either
add it in later when we eventually use this repository inside Gauss, or
move it to EvtGenExtras; prefer the former.

We may also not need EvtNoRadCorr.hh.

Added missing EvtHQET2FF.cpp bug fix for a normalisation term (f0m) that was
in the LHCb version but not copied over.