Fix apparent variable shadowing bug in EvtPropSLPole
Thanks to warning from clang8 that private field was unused, noticed
that three member variables were shadowed by local variables in one
member function, which appeared to be a mistake.
We should maybe consider enabling -Wshadow in future.
Need to check the impact of this bug fix on this model.