Rewrite EvtFlatSqDalitz model to be more efficient and to avoid cut-off around the edges
Summary:
With new version we have now also opportunity to generate rectangular part of the squared Dalitz plot with same efficiency.
Shuffle part of the code in EvtThreeBodyPhsp model which went from invariant masses squared to momenta of daughters to avoid duplication of the code and to allow reuse of it in EvtSqDalitz.
Improvement of EvtFlatSqDalitz model

Test Plan: Updated FlatSqDalitz model as well as changes to EvtThreeBodyPhsp were validated using test already in repository or new one added here.

Reviewers: tlatham, jback

Reviewed By: tlatham, jback

Tags: #evtgen

Differential Revision: https://phab.hepforge.org/D36