Allow to pass particle properties table in form of stringstream to constructor of EvtGen
Summary: In LHCb Gauss we are generating particle properties table on fly, which has to be written to the file in order to pass it to EvtGen constructor. This has the side effect that there is no good way of creating temporary file in safe way and current compilers do not like the function used to create temporary files as they are prone to race condition. In order to get rid of it, I added to EvtGen possibility to pass particle properties table in form of stringstream so we can eventually modify Gauss not to write temporary file at all or if needed write it in way which is not prone to race condition. Change here is fully transparent in a sense that existing code using EvtGen will work without any change.

Test Plan: I did run one of the tests with all changes and it works as expected.

Reviewers: tlatham, jback

Reviewed By: tlatham, jback

Tags: #evtgen

Differential Revision: https://phab.hepforge.org/D43