Model to generate 4-body phase-space decays in restricted part of the m12-m34 space.
Summary: The 4-body phase-space can be rather large and chance to obtain events in the corners of m12-m34 space is rather small. There are specific use cases where one wants to generate only events in the restricted range in m12-m34 plane and as one might want to select only small part or because small probability in the corners, just selecting events from full phase-space decays can be very inefficient. This model solves that by generating only events in wanted part of m12-m34 plane and sets maximum probability appropriately for the given window. Full phase-space can be generated with this model, but might not be best option for that.

Test Plan: In the process of testing, results of this decay model is compared to those from generic PHSP model including cases of restricted window.

Reviewers: tlatham, jback

Reviewed By: tlatham, jback

Tags: #evtgen

Differential Revision: https://phab.hepforge.org/D47